# BE/ecommerce-api/swagger.yaml
openapi: 3.0.3
info:
  title: E-commerce API
  description: |
    Comprehensive E-commerce REST API providing authentication, product management, shopping cart, and order processing capabilities.
    
    ## Authentication
    This API uses JWT tokens for authentication. Include the JWT token in the Authorization header as `Bearer <token>`.
    
    ## Base URL
    - Development: `http://localhost:8080`
    - Frontend Origins: `http://localhost:3000`, `http://localhost:3001`
    
    ## Features
    - User authentication and registration
    - Product catalog management with search and filtering
    - Shopping cart operations
    - Order management and tracking
    - User profile management
    - Admin functions for user and product management
  version: 1.0.0
  contact:
    name: E-commerce API Support
    email: support@ecommerce-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve all products with pagination, sorting, and filtering
      operationId: getAllProducts
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            default: "name"
            enum: ["name", "price", "category", "brand", "averageRating", "createdAt"]
        - name: sortDir
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            default: "asc"
            enum: ["asc", "desc"]
        - name: category
          in: query
          description: Filter by category
          required: false
          schema:
            type: string
        - name: subCategory
          in: query
          description: Filter by subcategory
          required: false
          schema:
            type: string
        - name: brand
          in: query
          description: Filter by brand
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
            format: decimal
        - name: maxPrice
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
            format: decimal
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductResponse'
                  pageable:
                    type: object
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer

    post:
      tags:
        - Products
      summary: Create new product
      description: Create a new product (Admin only)
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
            examples:
              new_product:
                summary: New product creation
                value:
                  sku: "PROD-ABC-123"
                  name: "Premium Wireless Headphones"
                  description: "Immersive sound, comfortable design, and long-lasting battery life"
                  category: "Electronics"
                  subCategory: "Headphones"
                  brand: "AudioTech"
                  price: 199.99
                  currency: "USD"
                  stockQuantity: 50
                  images: 
                    - "https://example.com/images/headphones_main.jpg"
                    - "https://example.com/images/headphones_side.jpg"
                  specifications:
                    connectivity: "Bluetooth 5.0"
                    batteryLifeHours: 20
                    noiseCancellation: true
                    colorOptions: ["Black", "Silver", "Blue"]
                  variations:
                    - sku: "PROD-ABC-123-BLK"
                      color: "Black"
                      price: 199.99
                      stockQuantity: 25
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a specific product by its ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found

    put:
      tags:
        - Products
      summary: Update product
      description: Update an existing product (Admin only)
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
        '401':
          description: Unauthorized

    delete:
      tags:
        - Products
      summary: Delete product
      description: Soft delete a product (Admin only)
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
      responses:
        '200':
          description: Product deleted successfully
        '404':
          description: Product not found
        '401':
          description: Unauthorized
  /products/sku/{sku}:
    get:
      tags:
        - Products
      summary: Get product by SKU
      description: Retrieve a specific product by its SKU
      operationId: getProductBySku
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
          example: "PROD-ABC-123"
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
  /products/category/{category}:
    get:
      tags:
        - Products
      summary: Get products by category
      description: Retrieve all products in a specific category
      operationId: getProductsByCategory
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
          example: "Electronics"
      responses:
        '200':
          description: Products in category retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /products/category/{category}/subcategory/{subCategory}:
    get:
      tags:
        - Products
      summary: Get products by subcategory
      description: Retrieve all products in a specific category and subcategory
      operationId: getProductsBySubCategory
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
          example: "Electronics"
        - name: subCategory
          in: path
          required: true
          schema:
            type: string
          example: "Headphones"
      responses:
        '200':
          description: Products in subcategory retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /products/brand/{brand}:
    get:
      tags:
        - Products
      summary: Get products by brand
      description: Retrieve all products from a specific brand
      operationId: getProductsByBrand
      parameters:
        - name: brand
          in: path
          required: true
          schema:
            type: string
          example: "AudioTech"
      responses:
        '200':
          description: Products from brand retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /products/search:
    get:
      tags:
        - Products
      summary: Search products
      description: Search products by name, description, brand, or category
      operationId: searchProducts
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          example: "wireless headphones"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            default: "name"
        - name: sortDir
          in: query
          required: false
          schema:
            type: string
            default: "asc"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
  /products/price-range:
    get:
      tags:
        - Products
      summary: Get products by price range
      description: Retrieve products within a specific price range
      operationId: getProductsByPriceRange
      parameters:
        - name: minPrice
          in: query
          required: true
          schema:
            type: number
            format: decimal
          example: 50.00
        - name: maxPrice
          in: query
          required: true
          schema:
            type: number
            format: decimal
          example: 500.00
      responses:
        '200':
          description: Products in price range retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
...
  /products/categories/{category}/subcategories:
    get:
      tags:
        - Products
      summary: Get subcategories by category
      description: Retrieve all subcategories for a specific category
      operationId: getSubCategoriesByCategory
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
          example: "Electronics"
      responses:
        '200':
          description: Subcategories retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["Headphones", "Smartphones", "Laptops", "Speakers"]
...
  /products/{id}/reviews:
    get:
      tags:
        - Products
      summary: Get product reviews
      description: Retrieve all reviews for a specific product
      operationId: getProductReviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            default: "date"
        - name: sortDir
          in: query
          required: false
          schema:
            type: string
            default: "desc"
      responses:
        '200':
          description: Product reviews retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductReviewResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer

    post:
      tags:
        - Products
      summary: Add product review
      description: Add a review for a specific product
      operationId: addProductReview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductReviewRequest'
            examples:
              new_review:
                summary: New product review
                value:
                  productId: "654321abcdef01234567890"
                  userId: "987654fedcba01234567890"
                  rating: 5
                  comment: "Amazing sound quality!"
                  isVerifiedPurchase: true
      responses:
        '201':
          description: Review added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductReviewResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
  /products/{id}/rating-summary:
    get:
      tags:
        - Products
      summary: Get product rating summary
      description: Get rating statistics for a specific product
      operationId: getProductRatingSummary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
      responses:
        '200':
          description: Rating summary retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  averageRating:
                    type: number
                    format: double
                    example: 4.5
                  totalReviews:
                    type: integer
                    example: 25
                  ratingDistribution:
                    type: object
                    properties:
                      "1":
                        type: integer
                        example: 1
                      "2":
                        type: integer
                        example: 2
                      "3":
                        type: integer
                        example: 3
                      "4":
                        type: integer
                        example: 7
                      "5":
                        type: integer
                        example: 12
  /products/{id}/stock:
    put:
      tags:
        - Products
      summary: Update product stock
      description: Update the stock quantity for a product (Admin only)
      operationId: updateProductStock
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stockQuantity
              properties:
                stockQuantity:
                  type: integer
                  minimum: 0
            example:
              stockQuantity: 75
      responses:
        '200':
          description: Stock updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
        '401':
          description: Unauthorized
  /products/{id}/variations/{variationSku}/stock:
    put:
      tags:
        - Products
      summary: Update variation stock
      description: Update the stock quantity for a product variation (Admin only)
      operationId: updateVariationStock
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "654321abcdef01234567890"
        - name: variationSku
          in: path
          required: true
          schema:
            type: string
          example: "PROD-ABC-123-BLK"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stockQuantity
              properties:
                stockQuantity:
                  type: integer
                  minimum: 0
            example:
              stockQuantity: 30
      responses:
        '200':
          description: Variation stock updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product or variation not found
        '401':
          description: Unauthorized
  /products/featured:
    get:
      tags:
        - Products
      summary: Get featured products
      description: Retrieve featured products (high-rated and popular items)
      operationId: getFeaturedProducts
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Featured products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /products/low-stock:
    get:
      tags:
        - Products
      summary: Get low stock products
      description: Retrieve products with low stock quantities (Admin only)
      operationId: getLowStockProducts
      security:
        - bearerAuth: []
      parameters:
        - name: threshold
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Low stock products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        '401':
          description: Unauthorized
  /products/top-rated:
    get:
      tags:
        - Products
      summary: Get top-rated products
      description: Retrieve products with highest ratings
      operationId: getTopRatedProducts
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - name: minRating
          in: query
          required: false
          schema:
            type: number
            format: double
            default: 4.0
            minimum: 1.0
            maximum: 5.0
      responses:
        '200':
          description: Top-rated products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

  # Cart Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. Include as 'Bearer <token>'
  schemas:
...
    ProductRequest:
      type: object
      required:
        - sku
        - name
        - price
        - stockQuantity
      properties:
        sku:
          type: string
          maxLength: 50
          description: Product SKU (Stock Keeping Unit)
          example: "PROD-ABC-123"
        name:
          type: string
          maxLength: 100
          description: Product name
          example: "Premium Wireless Headphones"
        description:
          type: string
          maxLength: 1000
          description: Product description
          example: "Immersive sound, comfortable design, and long-lasting battery life"
        category:
          type: string
          maxLength: 50
          description: Product category
          example: "Electronics"
        subCategory:
          type: string
          maxLength: 50
          description: Product subcategory
          example: "Headphones"
        brand:
          type: string
          maxLength: 50
          description: Product brand
          example: "AudioTech"
        price:
          type: number
          format: decimal
          minimum: 0.01
          description: Product price
          example: 199.99
        currency:
          type: string
          maxLength: 3
          description: Currency code
          example: "USD"
        stockQuantity:
          type: integer
          minimum: 0
          description: Available stock quantity
          example: 50
        images:
          type: array
          items:
            type: string
            format: uri
          description: Product image URLs
          example: 
            - "https://example.com/images/headphones_main.jpg"
            - "https://example.com/images/headphones_side.jpg"
        specifications:
          $ref: '#/components/schemas/ProductSpecifications'
        variations:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariation'
        isActive:
          type: boolean
          description: Whether the product is active
          example: true
...